Phase_III-(todo-chatbot)/backend/requirements.txt
fastapi[all]==0.104.1
sqlmodel==0.0.16
psycopg[binary]==3.3.2
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
uvicorn[standard]==0.24.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
mcp==0.9.0
openai==1.54.0
python-dotenv==1.0.0
